+++ CHANGELOG.md
@@ -14,11 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - The Agiltron FF optical switch QMI drivers have now common base class in `qmi.instruments.agiltron._ff_optical_switch`
 - Refactored also unit-tests for Agiltron FF optical switches.
+- Zurich Instruments HDAWG instrument driver now uses `zhinst.core` instead of `zhinst.ziPython`. Also the 'schema' is obtained now from the instrument itself, and not from a separate file.
 
 ### Fixed
 - In `usbtmc.py` now doing `.strip()` on `dev.serial_number` string to avoid SNs with whitespace character(s).
 
 ### Removed
+- Zurich Instruments HDAWG instrument driver `hdawg_command_table.schema` file.
 
 ## [0.50.0] - 2025-09-01